frontend/requirements.txt
streamlit
pillow
pandas
pathlib
plotly
os
streamlit-ace
numpy
sys